test refactor ui
